backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
langchain==0.0.335
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.2
httpx==0.25.2
pdfplumber==0.10.3
pandas==2.1.3
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1 